requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.5
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cloudpathlib==0.22.0
confection==0.1.5
cryptography==46.0.1
cymem==2.0.11
dataclasses-json==0.6.7
distro==1.9.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.19.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
fsspec==2025.9.0
fuzzywuzzy==0.18.0
greenlet==3.2.4
groq==0.31.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-pinecone==0.2.12
langchain-text-splitters==0.3.11
langcodes==3.5.0
langsmith==0.4.31
language_data==1.3.0
Levenshtein==0.27.1
Mako==1.3.10
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
murmurhash==1.0.13
mypy_extensions==1.1.0
mysql-connector-python==9.4.0
mysqlclient==2.2.7
networkx==3.5
numpy==2.3.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.109.1
orjson==3.11.3
packaging==24.2
pandas==2.3.2
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
preshed==3.0.10
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.10
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.27.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.1
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
simsimd==6.5.3
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
soupsieve==2.8
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
torch==2.8.0
tqdm==4.67.1
transformers==4.56.2
triton==3.4.0
typer==0.19.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
zstandard==0.25.0
